drivers: adc/mspm0: add power management for vref
when ADC channel is setup with using internal vref, vref IP is enabled
and constantly in ON state. This consumes ~28uA even when ADC is not in
use.

Best approach is to split the ADC driver into two parts,
- regulator driver for Vref
- ADC driver depending on Vref over phandle
as these changes will be involved, we will move this upstream planning

With the current changes, vref will be enabled / disabled based on PM
device runtime API's to reduce the power usage.

Signed-off-by: Parthiban Nallathambi <parthiban@linumiz.com>