soc: mspm0g: override sys_arch_reboot() to trigger power-on reset
the default implementation of sys_reboot(SYS_REBOOT_COLD) performs a
CPU-only reset. This commit overrides the weak ARM implementation
with a device-specific function for the MSPM0G series that triggers
a power-on reset instead.
This ensures a more complete reset of the device when SYS_REBOOT_COLD
is requested

Signed-off-by: [Jakob Schulz] <jschulz@testo.de>