drivers: ethernet: adapt cadence for traveo
Existing works in upstream and outstanding PR [1] are primarily meant
for AMD / Xilinx platform MPU in 64 bit systems. With traveo, following
primary differences applies,
- 32 bit M7 platform
- DMA is interfaced using AMBA AXI instead of AHB. Also to add, AXI data
bus width is 64 bit compared with AHB's 32bit.
- SoC specific register definitions varies. For example, 64 bit platform
splits the queue base address pointer as upper and lower, whereas with
32 bit T2G, it will be handled differently as separate queue used with QoS
or priority
- Clock inputs for the EMAC varies based on PHY link speeds
- Pin controller setup needs to be done
- MDIO clock and pin controller varies

So to accomodate all the differences, all the changes are clubbed into
one commit in this history. Ideal idea is to split for upstream next.

Signed-off-by: Karthikeyan Krishnasamy <karthikeyan@linumiz.com>