dts: infineon: remove CM0+ image configuration and linker scripts
This commit removes the configuration options and linker scripts related
to the CM0+ image for the CYT4DN series. The changes include the
deletion of the CM0+ image RAM and ROM size configurations, as well
as the corresponding linker script files. This cleanup is necessary
to streamline the build process and eliminate unused configurations.

Signed-off-by: Christoph Seitz <christoph.seitz@infineon.com>