soc: infineon: add Kconfig support for M0P launcher and sysbuild integration
This commit introduces Kconfig support for the M0P
launcher image based on the minimal example in the
Infineon SOC. It adds necessary configurations in the Kconfig files, ensuring that the M0P launcher can be
included in the build process. Additionally, a CMake file is created to
handle the sysbuild process for the M0P launcher, setting up the required
parameters for the build.

These changes are essential for enabling the M0P image integration
into M7 builds to access SROM APIs.

Signed-off-by: Christoph Seitz <christoph.seitz@infineon.com>